vkernel64: Introduce checkpoint signal handler
Special care must be taken in order not to corrupt physical memory
backing file which will be used by the vkernel after checkpoint restore.
Signal handler uses alternative stack which is not mapped using
MAP_VPAGETABLE. In addition to that, vkernel is not allowed to continue
execution after saving the checkpoint file.